Backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "studhelper-backend"
version = "2.0.0"
description = "AI-powered study assistant backend"
authors = [{name = "StudHelper Team"}]
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "sqlalchemy>=2.0.23",
    "alembic>=1.13.1",
    "asyncpg>=0.29.0",
    "python-multipart>=0.0.6",
    "openai>=1.6.1",
    "tiktoken>=0.5.2",
    "tenacity>=8.2.3",
    "pdfplumber>=0.10.3",
    "pypdf>=3.17.4",
    "python-docx>=1.1.0",
    "python-pptx>=0.6.23",
    "yt-dlp>=2023.12.30",
    "openai-whisper>=20231117",
    "chromadb>=0.4.22",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.2",
    "langchain-chroma>=0.1.0",
    "sentence-transformers>=2.2.2",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "httpx>=0.25.2",
    "aiofiles>=23.2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "black>=23.12.1",
    "flake8>=6.1.0",
    "mypy>=1.8.0",
    "isort>=5.13.2",
]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"

